compiles!

+++ contracts/governance/GovTokenTypes.sol
@@ -5,5 +5,5 @@ pragma solidity ^0.8.0;
 /// @notice A checkpoint for marking number of votes from a given block
 struct Checkpoint {
     uint32 fromBlock;
-    uint96 votes;
+    uint votes;
 }
